PublicKey::verify_signature constant time
The existential unforgeability under a single-message attack of Lamport's
One-Time Signature Scheme does not depend on at all on the constant-time
verification of a signature.

Not being able to anticipate all threat models, there may be some in which some
signed data or its signature is intended to be private, and an adversary who can
time the verification of such data may learn something.