Refactor PublicKey::to_bytes
Use of the specialized `extend_from_slice` method vs. `append`, and also the
explicit setting of the capacity, so reallocation on the heap is never required,
both improved the performance of this function.

Add benchmark for to_bytes

This requires we remove the unstable_features lint, but does not require we add
any unstable Rust features to the library code--only to the tests.